+++ qml/ControllerDialog.qml
@@ -63,9 +63,15 @@ WalletDialog {
                 return
             }
             if (resolver instanceof SignLiquidTransactionResolver) {
-                stack_view.push(sign_liquid_transaction_resolver_view_component, { resolver, wallet })
-                resolver.resolve()
-                return
+                if (resolver.device instanceof JadeDevice) {
+                    const view = jade_sign_liquid_transaction_view.createObject(stack_view, { resolver, wallet })
+                    stack_view.push(view)
+                    resolver.resolve()
+                } else {
+                    stack_view.push(sign_liquid_transaction_resolver_view_component, { resolver, wallet })
+                    resolver.resolve()
+                }
+                return                    
             }
             if (resolver instanceof SignMessageResolver) {
                 if (resolver.device instanceof JadeDevice) {
@@ -91,6 +97,11 @@ WalletDialog {
     Component {
         id: sign_liquid_transaction_resolver_view_component
         SignLiquidTransactionResolverView {}
+    }
+
+    Component {
+        id: jade_sign_liquid_transaction_view
+        JadeSignLiquidTransactionView {}
     }
 
     Component {