+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Show Tor indicator in status bar when Tor is active
+- Show liquid transaction details when signing with Jade
 
 ### Changed
 - New icons for Liquid and Liquid testnet